+++ gwcelery/tasks/superevents.py
@@ -4,7 +4,7 @@ superevents.
 *   There is serial processing of triggers from low latency pipelines.
 *   Dedicated **superevent** queue for this purpose.
 *   Primary logic to respond to low latency triggers contained in
-    :meth:`handle` function.
+    :meth:`process` function.
 """
 from celery.utils.log import get_task_logger
 from ligo.segments import segment, segmentlist
@@ -21,13 +21,12 @@ log = get_task_logger(__name__)
                  'cbc_mbtaonline',
                  'burst_olib',
                  'burst_cwb',
-                 queue='superevent',
                  shared=False)
 def handle(payload):
-    """LVAlert handler for superevent manager.
-
-    Receives payload from test and production nodes and serially processes them
-    to create/modify superevents."""
+    """Respond to lvalert nodes from low-latency search
+    pipelines and delegate to :meth:`process` for
+    superevent management.
+    """
     if payload['alert_type'] != 'new':
         return
 
@@ -41,10 +40,26 @@ def handle(payload):
         return
     else:
         if far > app.conf['superevent_far_threshold']:
-            log.info("Skipping processing of %s because of low FAR", gid)
+            log.info("Skipping processing of %s because of high FAR", gid)
             return
+    process.delay(payload)
 
+
+@gracedb.task(queue='superevent', shared=False)
+@gracedb.catch_retryable_http_errors
+def process(payload):
+    """
+    Respond to `payload` and serially processes them
+    to create new superevents, add events to existing ones
+    and update superevent parameters.
+
+    Parameters
+    ----------
+    payload : dict
+        LVAlert payload
+    """
     event_info = payload['object']
+    gid = payload['uid']
 
     if event_info.get('search') == 'MDC':
         category = 'mdc'
@@ -68,7 +83,6 @@ def handle(payload):
     t_0, t_start, t_end = get_ts(event_info)
 
     if sid is None:
-        log.debug('Entered 1st if')
         event_segment = _Event(event_info['gpstime'],
                                t_start, t_end,
                                event_info['graceid'],